Create CODE_OF_CONDUCT.md
No matter how upset you may be, or how frustrated you may be, as a community we have standards of behaviour that we either accept or do not accept.

If you ever feel threatened by another user or feel material posted on our forum is inappropriate, please do your part in showing the kind of material and behaviour you find unacceptable by using the 'report to a moderator link which can be found next to each post. This is not about censorship, this is about maintaining an environment where people feel safe and willing to contribute and not have to be treated poorly by others. Even if you have something difficult or contradictory to say, you can still say it in a non-confrontational and respectful manner.

Thanks to the 99.9% of our users who act like the mature people they are, we can rise above this poor behaviour and continue on doing our best as an open-source community. 

Other projects and sites may accept poor behaviour, we do not!